refactor: extract page content and sidebar components
